feat(BigQuery):The target is written in two modes.
The target is written in two modes: batch mode (stream api) and stream mode (stream temporary table+merge to target table)
When the target only receives the insert event, it will write in batch mode
When the target receives the first update/delete event, it will be written in stream mode
The time for the first merge is at least 31 minutes, and the subsequent merge time is configured according to the user's configuration
After each merge, it will return to batch mode, change to stream mode after receiving U/D, and record this time T1 (save to stateMap at the same time). The next merge will use T1+the time configured by the user as the merge time point.
When the task is stopped and restarted, set the next merge time according to T1. If the current time is greater than T1+the time interval configured by the user, merge immediately.