Feat connectors bigquery (#400)
* feat(BigQuery):The target is written in two modes.

The target is written in two modes: batch mode (stream api) and stream mode (stream temporary table+merge to target table)
When the target only receives the insert event, it will write in batch mode
When the target receives the first update/delete event, it will be written in stream mode
The time for the first merge is at least 31 minutes, and the subsequent merge time is configured according to the user's configuration
After each merge, it will return to batch mode, change to stream mode after receiving U/D, and record this time T1 (save to stateMap at the same time). The next merge will use T1+the time configured by the user as the merge time point.
When the task is stopped and restarted, set the next merge time according to T1. If the current time is greater than T1+the time interval configured by the user, merge immediately.

* fix(BigQuery): After creating a temporary table, reset needs to delete the temporary table .

* fix(BigQuery): Data insertion exception in stream mode and improvement of merge SQL .

* fix(BigQuery): Add 2 minutes delay for the first merge .

* fix(BigQuery):close stream.

* fix:close stream after append over.

* fix(BigQuery):shut down client and await termination before close stream.

* fix(BigQuery): formatting code .

* fix(BigQuery):close client and flush rows.

* fix(BigQuery):change log level.

* fix(BigQuery):temporary event data, set insert event before data.

* fix(BigQuery):temporary event data, set insert event before data.