fix(lark-js-connector):Fly book as the target to send messages.
A field receiveType, which refers to the type of message receiver, mainly group chat or user, can be added to the format of the data written by the flying book as the target;

In this way, the corresponding receiveId can be controlled according to the receiveType,

If receiveType=user, receiveId is transferred to the user's open_ id；

If receiveType=email, the receiveId is transferred to the user's email;

If receiveType=phone, the receiveId is transferred to the user's phone. At this time, the user's open is obtained through the phone drop API_ id；

If receiveType=chat, receiveId will send the chat to the group_ id, chat_ The id needs to be processed at the processor node;

At the same time, add a contentType attribute to indicate the type of message sent, which can be text (text), interactive (card), post (rich text), image (image)

In this case, the data source focuses on the action of calling API, and the construction of this record structure is done by the processor node.